backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
chromadb
langchain
requests
pypdf
python-docx 